requirements.txt
tomli >= 1.1.0 ; python_version < '3.11'
numpy ~= 1.23.5
scipy
tensorflow ~=2.8
matplotlib
